Move functions and states into useBooks hook
